refactor: move manual scanning workflow to application
