+++ tests/architecture/test_cleanup_ratchet.py
@@ -25,7 +25,6 @@ TEMPORARY_APPLICATION_TO_INFRASTRUCTURE = {
     ("application/runtime_builder.py", "anteumbra.infrastructure.monitoring.log_analyzer"),
     ("application/runtime_builder.py", "anteumbra.infrastructure.monitoring.log_monitor"),
     ("application/runtime_builder.py", "anteumbra.infrastructure.monitoring.monitor"),
-    ("application/scanner_service.py", "anteumbra.infrastructure.detection.manual_scanner"),
     ("application/yara_service.py", "anteumbra.infrastructure.detection.yara_engine"),
 }
 
@@ -159,7 +158,6 @@ TEMPORARY_APPLICATION_TO_INFRASTRUCTURE_RATIONALES = {
     ("application/runtime_builder.py", "anteumbra.infrastructure.monitoring.log_analyzer"): "inject worker analyzer factory",
     ("application/runtime_builder.py", "anteumbra.infrastructure.monitoring.log_monitor"): "inject log monitor factory",
     ("application/runtime_builder.py", "anteumbra.infrastructure.monitoring.monitor"): "inject site monitor factory",
-    ("application/scanner_service.py", "anteumbra.infrastructure.detection.manual_scanner"): "define scan port",
     ("application/yara_service.py", "anteumbra.infrastructure.detection.yara_engine"): "define YARA scan port",
 }
